logic/app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
mysql-connector-python
Jinja2
httpx